[PVParallelView] fix lose of some shortcuts when many views are opened
if you open ascatter view and a hit-count view, the shortcuts which are
common to the 2 views do not work anymore.

This bug exists in all 4.x versions (maybe in 3.x:)

As the views are in the same top-level window, the QAction' event filter
fails to identify the target parent widget.

Using the shortcut context Qt::WidgetWithChildrenShortcut instead of
Qt::WindowShortcut (the defautlt) solves that problem.