[splitter_key_value] rework listwidget buttons enabling/disabling
move up/down, copy, delete buttons are always enabled even if their actions
are not possible.

They are now enable (or not) according to the list state.

In case of multiple selection, the move up/down may not work because the
selection list order is important for the used algorithms.

Empty selection check before use has been added too. close #524