[converter_substitution] rework tablewidget buttons enabling/disabling
move up/down, copy, delete buttons are always enabled even if their actions
are not possible.

They are now enable (or not) according to the list state.

In case of selection move up or down, the selection was always changed to the
last moved row. This is due to the internal behaviour of the selection mode
ExtendedSelection. Using temporarily the MultiSelection mode help keeping the
current selection unchanged.