[GuiQt] fix uncaught exception when there is no format string for the time type
In case of empty string for the time type format, the exception
PVFormatNoTimeMapping was thrown but never caught. As it complained about a
mapping problem and not about a format parameter problem, the exception
PVFormatInvalidTime has been introduced.

However, inspector can still crash, even after this exception has been caught.
This is due to an undefined state of the _nraw member variable. Doing some
resets in case of exception solves this second crash.

Minor style harmonization for the PVRush::PVInputException case.

close #523