[PVParallelView] fix gray zones when changing the axes combination
When there are more than PARALLELVIEW_MAX_DRAWN_ZONES axes, removing one axis
under this limit makes the 29th zone weird: the background is gray and the
selection is empty.

This problem appends because a non-displayed zone becomes implictly displayable.

To solve this problem, all the zones are invalidated to make them correct. The
reasons are:
* all other cases of zones invalidation process all zones, whatever they are
  displayable or not;
* detecting the zones to make displayable needs to check for each full parallel
  views if the axes combination change can impact the zones displayability or
  not, adding complex algorithm and/or doing many code changes.

The list of new zones to create is now useless outside of
PVZonesManager::update_from_axes_comb(...)

close #476