[PVGuiQt] fix and simplify sort in stat-views
When clicking for the first time on the values column, the values are
correctly sorted but the sort indicator is wrong. Clicking a second time does
not change the order but changes the indicator. Clicking a third time leads
to a coherent time.

Strangly, this bug does not affect the stat columns.

This bug can be fixed (and the code simplified) by using QTableView's column
sorting logic instead of rewriting a new one.

Defining a different initial sort order according to the column can be done
by handling the role Qt::InitialSortOrderRole in the model headerData()
method.

Qt signal layoutAboutToBeChanged must be emitted by a model before it starts
sorting its data.