[Inendi] permit to set a parent widget to the "new layer" dialog
The "new layer" dialog has no parent by default, changing mouse focus also
makes this dialog dissappear.

In issue #577, the steps to reproduce the problem moves a dockable view
outside of the main window which can over the main window or on an other
screen. Using the window under the mouse instead of main window permits to
have the "new layer" dialog near to the mouse instead of on an other screen.